+++ Code/Examples/examples/src/main/scala/examples/demo/KSemiModularBall.scala
@@ -39,7 +39,7 @@ object KSemiModularBall extends Main {
     val reset = resetIn.map(pos => Left[Point, Pos](pos))
 
     val pos = (reset || inc).fold(Pos(0,0)){
-      case (_, Left(Point(x, y))) => Pos(x, y)
+      case (_, Left(Point(x, y))) => Pos(x.toDouble, y.toDouble)
       case (pX, Right(inc)) => pX + inc
     }
